+++ src/main/java/org/vniizht/suburbsweb/service/data/entities/level3/co22/T6.java
@@ -12,9 +12,9 @@ import lombok.ToString;
 public class T6 {
     private String          p1;
     private String          p2;
-    private Integer         p3;
+    private Long            p3;
     private Short           p4;
     private String          p5;
-    private Integer         p6;
+    private Short           p6;
     private Short           p7;
 }
